examples/cpp-grpc-fastapi-example/py-client/requirements.txt
fastapi
uvicorn[standard]
grpcio
grpcio-tools
protobuf
